file checks and run through all genes - lower threshold for considering alleles

+++ scripts/filter_immune_subgraph.sh
@@ -92,33 +92,40 @@ sample_immune_aln=${sample}.${graph}.${aln_type}
 immune_graph=${graph_base}.subgraph
 cd ${workdir}
 #
-echo "Embedding sample-specific variation into the whole immune subgraph"
 output_graph=${sample_id}_graph.${graph}.${aln_type}
-# More conservative filtering at the gene level based on vg map - v liberal thresh here using pairwise identity to locus + graph augmentation using reads with base quality > 5 and breakpoints with coverage > 2
-vg filter -r 0.9 -s 1 -fu -t 4 ${sample_immune_graph}.gam -D 0 -v > ${sample_immune_graph}.filtered.gam
-vg augment -s -m 2 -q 5 ${immune_graph}.vg ${sample_immune_graph}.filtered.gam > ${output_graph}.vg
-echo "normalizing, condensing, sorting augmented graph"
-vg mod ${output_graph}.vg -n -U 100 -c -X 256 - | vg ids -c - > ${sample_id}.tmp && mv ${sample_id}.tmp ${output_graph}.vg
-vg convert -fW ${output_graph}.vg > ${output_graph}.gfa
-odgi build -g ${output_graph}.gfa -OP -o ${output_graph}.og
-odgi normalize -i ${output_graph}.og -o ${output_graph}.norm.og -P
-#odgi sort -i ${output_graph}.norm.og -p Ygs -P -O -o ${output_graph}.sort.og
-odgi view -i ${output_graph}.norm.og -g -a -P >${output_graph}.norm.gfa
-gfaffix ${output_graph}.norm.gfa -o ${output_graph}.fix.gfa
+if [ -s ${output_graph}.fix.gfa ]; then
+    echo "Alignment completed - using: ${sample}.${graph_base##*/}.gam";
+else
+    echo "Embedding sample-specific variation into the whole immune subgraph"
+    # More conservative filtering at the gene level based on vg map - v liberal thresh here using pairwise identity to locus + graph augmentation using reads with base quality > 5 and breakpoints with coverage > 2
+    vg filter -r 0.9 -s 1 -fu -t 4 ${sample_immune_graph}.gam -D 0 -v > ${sample_immune_graph}.filtered.gam
+    vg augment -s -m 2 -q 5 ${immune_graph}.vg ${sample_immune_graph}.filtered.gam > ${output_graph}.vg
+    echo "normalizing, condensing, sorting augmented graph"
+    vg mod ${output_graph}.vg -n -U 100 -c -X 256 - | vg ids -c - > ${sample_id}.tmp && mv ${sample_id}.tmp ${output_graph}.vg
+    vg convert -fW ${output_graph}.vg > ${output_graph}.gfa
+    odgi build -g ${output_graph}.gfa -OP -o ${output_graph}.og
+    odgi normalize -i ${output_graph}.og -o ${output_graph}.norm.og -P
+    #odgi sort -i ${output_graph}.norm.og -p Ygs -P -O -o ${output_graph}.sort.og
+    odgi view -i ${output_graph}.norm.og -g -a -P >${output_graph}.norm.gfa
+    gfaffix ${output_graph}.norm.gfa -o ${output_graph}.fix.gfa
+fi
 #
 # cleaning assumes ngmerged route takes sufficiently longer - pe workflow has completed
 if [[ ${aln_type} == *"pe"* ]]; then
 # perform some depth estimates:
-    vg depth --gam ${sample_immune_graph}.gam ${immune_graph}.xg > ${sample}.${graph}_immune.pe.depth;
-    depth_immune=$(awk -F ' ' '{print $1}' ${sample}.${graph}_immune.pe.depth)
-    echo "${sample} depth across immunovariation loci: ${depth_immune}"
-    # TTN exon 327 depth - immunotyper-SR reference
-    #chr2:179424038-179441143
-    grch_path_chr2=$(vg paths -Lv ${graph_base}.xg | grep "grch38#chr2:")
-    vg depth --gam ${sample_immune_graph}.gam ${graph_base}.xg -p "${grch_path_chr2}:179424038[-179441143]" > ${sample}.${graph}_TTN.pe.depth;
-    depth_ttn=$(awk -F ' ' '{print $1}' ${sample}.${graph}_TTN.pe.depth)
-    echo "${sample} depth across Titin (TTN) exon 327 (Immunotyper-sr ref): ${depth_ttn}"
-# finished!
+    if [ -s ${sample}.${graph}_TTN.pe.depth ]; then
+        echo "depth estimates exist - using them"
+    else
+        vg depth --gam ${sample_immune_graph}.gam ${immune_graph}.xg > ${sample}.${graph}_immune.pe.depth;
+        depth_immune=$(awk -F ' ' '{print $1}' ${sample}.${graph}_immune.pe.depth)
+        echo "${sample} depth across immunovariation loci: ${depth_immune}"
+        # TTN exon 327 depth - immunotyper-SR reference
+        #chr2:179424038-179441143
+        grch_path_chr2=$(vg paths -Lv ${graph_base}.xg | grep "grch38#chr2:")
+        vg depth --gam ${sample_immune_graph}.gam ${graph_base}.xg -p "${grch_path_chr2}:179424038[-179441143]" > ${sample}.${graph}_TTN.pe.depth;
+        depth_ttn=$(awk -F ' ' '{print $1}' ${sample}.${graph}_TTN.pe.depth)
+        echo "${sample} depth across Titin (TTN) exon 327 (Immunotyper-sr ref): ${depth_ttn}"
+    fi
     echo "fin";
     #sbatch --export=i=${sample_immune_graph}.gam,sample_id=$sample,graph=$graph,aln_type='pe',use_augmented=FALSE,workdir=${PWD} /home/dd392/project/gene_hap_IMGT_vgflow.sh
     export i=${sample_immune_graph}.gam workdir=${PWD} sample_id=${sample} aln_type='pe' tools_dir=${tools_dir} use_augmented=FALSE bigfoot_source=${bigfoot_source} bigfoot_dir=${bigfoot_dir} merged=FALSE graph=${graph} valid_alleles=${valid_alleles}
